Feature server validation tl

+++ server/routes/events.js
@@ -5,21 +5,25 @@ var pool = require('../modules/db');
 
 router.get('/:session', function(req, res) {
   var session = req.params.session;
-  pool.connect(function(errorConnectingToDb, db, done) {
-    if (errorConnectingToDb) {
-      res.sendStatus(500);
-    } else {
-      db.query('SELECT * from "forms" JOIN "events" ON "events"."form_id" = "forms"."id" WHERE "session_id" = $1 ORDER BY "events"."meeting_count" ASC;', [session],
-        function(queryError, result) {
-          done();
-          if (queryError) {
-            res.sendStatus(500);
-          } else {
-            res.send(result.rows);
-          }
-        });
-    }
-  });
+  if (req.isAuthenticated()) {
+    pool.connect(function(errorConnectingToDb, db, done) {
+      if (errorConnectingToDb) {
+        res.sendStatus(500);
+      } else {
+        db.query('SELECT * from "forms" JOIN "events" ON "events"."form_id" = "forms"."id" WHERE "session_id" = $1 ORDER BY "events"."meeting_count" ASC;', [session],
+          function(queryError, result) {
+            done();
+            if (queryError) {
+              res.sendStatus(500);
+            } else {
+              res.send(result.rows);
+            }
+          });
+      }
+    });
+  } else {
+    res.sendStatus(401);
+  }
 }); //end router.get
 
 router.post('/add', function(req, res) {
@@ -28,22 +32,26 @@ router.post('/add', function(req, res) {
   var form_id = req.body.form_id;
   var open_date = req.body.open_date;
   var close_date = req.body.close_date;
-  pool.connect(function(errorConnectingToDb, db, done) {
-    if (errorConnectingToDb) {
-      res.sendStatus(500);
-    } else {
-      db.query('INSERT INTO "events" ("session_id", "meeting_count", "form_id", "date_form_open", "date_form_close") VALUES ($1,$2,$3,$4,$5);',
-      [session_id, meeting_count, form_id, open_date, close_date],
-        function(queryError, result) {
-          done();
-          if (queryError) {
-            res.sendStatus(500);
-          } else {
-            res.sendStatus(201);
-          }
-        });
-    }
-  });
+  if (req.isAuthenticated()) {
+    pool.connect(function(errorConnectingToDb, db, done) {
+      if (errorConnectingToDb) {
+        res.sendStatus(500);
+      } else {
+        db.query('INSERT INTO "events" ("session_id", "meeting_count", "form_id", "date_form_open", "date_form_close") VALUES ($1,$2,$3,$4,$5);',
+        [session_id, meeting_count, form_id, open_date, close_date],
+          function(queryError, result) {
+            done();
+            if (queryError) {
+              res.sendStatus(500);
+            } else {
+              res.sendStatus(201);
+            }
+          });
+      }
+    });
+  } else {
+    res.sendStatus(401);
+  }
 }); //end router.post
 
 router.put('/update', function(req, res) {
@@ -52,42 +60,50 @@ router.put('/update', function(req, res) {
   var form_id = req.body.form_id;
   var open_date = req.body.open_date;
   var close_date = req.body.close_date;
-  pool.connect(function(errorConnectingToDb, db, done) {
-  if (errorConnectingToDb) {
-    res.sendStatus(500);
+  if (req.isAuthenticated()) {
+    pool.connect(function(errorConnectingToDb, db, done) {
+    if (errorConnectingToDb) {
+      res.sendStatus(500);
+      } else {
+        db.query('UPDATE "events" SET "meeting_count"=$1, "form_id"=$2, "date_form_open"=$3, "date_form_close"=$4 WHERE "id" = $5;',
+        [meeting_count, form_id, open_date, close_date, id],
+          function(queryError, result) {
+            done();
+            if (queryError) {
+              res.sendStatus(500);
+            } else {
+              res.sendStatus(201);
+            }
+          });
+      }
+    });
   } else {
-    db.query('UPDATE "events" SET "meeting_count"=$1, "form_id"=$2, "date_form_open"=$3, "date_form_close"=$4 WHERE "id" = $5;',
-    [meeting_count, form_id, open_date, close_date, id],
-      function(queryError, result) {
-        done();
-        if (queryError) {
-          res.sendStatus(500);
-        } else {
-          res.sendStatus(201);
-        }
-      });
+      res.sendStatus(401);
   }
-});
 });//end router.put
 
 
 router.delete('/delete/:id', function(req, res) {
   var eventID = req.params.id;
-  pool.connect(function(errorConnectingToDb, db, done) {
-    if (errorConnectingToDb) {
-      res.sendStatus(500);
-    } else {
-      db.query('DELETE FROM "events" WHERE "id" = $1;', [eventID],
-        function(queryError, result) {
-          done();
-          if (queryError) {
-            res.sendStatus(500);
-          } else {
-            res.sendStatus(201);
-          }
-        });
-    }
-  });
+  if (req.isAuthenticated()) {
+    pool.connect(function(errorConnectingToDb, db, done) {
+      if (errorConnectingToDb) {
+        res.sendStatus(500);
+      } else {
+        db.query('DELETE FROM "events" WHERE "id" = $1;', [eventID],
+          function(queryError, result) {
+            done();
+            if (queryError) {
+              res.sendStatus(500);
+            } else {
+              res.sendStatus(201);
+            }
+          });
+      }
+    });
+  } else {
+       res.sendStatus(401);
+  }
 }); //end router.delete
